Create 5-functional-programming.md
Note - this file abides by the old order convention, in which there's an expose on functional programming at the beginning of the document. @parthsarin - I'm interested to hear your thoughts on this. I think there are valid points to be made against its inclusion (it's certainly less of a Python topic and more of a generic programming topic), but I'd also suggest we include it in the notes - though not the lecture - as an interesting way of setting up and motivating the material.